Ajustando comportamentos

+++ src/app/Http/Controllers/Web/InviteController.php
@@ -25,10 +25,19 @@ class InviteController extends Controller
         $this->authorizeResource(Invite::class, 'invite');
     }
 
-    public function index()
+    public function index(Request $request)
     {
-        $invites = $this->service->all();
-        return view('invites.index', compact('invites'));
+        $surveys = $this->surveyService->all();
+        $groups  = $this->groupService->all();
+        $query   = Invite::with(['survey', 'group']);
+        if ($request->filled('survey_id')) {
+            $query->where('survey_id', $request->survey_id);
+        }
+        if ($request->filled('group_id')) {
+            $query->where('group_id', $request->group_id);
+        }
+        $invites = $query->get();
+        return view('invites.index', compact('invites', 'surveys', 'groups'));
     }
 
     public function create()
@@ -44,35 +53,62 @@ class InviteController extends Controller
         return redirect()->route('invites.index');
     }
 
-    public function show(int $id)
+    public function show(Invite $invite)
     {
-        $invite = $this->service->find($id);
+        $this->authorize('view', $invite);
+        $invite = $this->service->find($invite->id);
         return view('invites.show', compact('invite'));
     }
 
-    public function edit(int $id)
+    public function edit(Invite $invite)
     {
-        $invite = $this->service->find($id);
+        $this->authorize('update', $invite);
+        $invite = $this->service->find($invite->id);
         $surveys = $this->surveyService->all();
         $groups = $this->groupService->all();
         return view('invites.edit', compact('invite', 'surveys', 'groups'));
     }
 
-    public function update(StoreInviteRequest $request, int $id)
+    public function update(StoreInviteRequest $request, Invite $invite)
     {
-        $this->service->update($id, $request->validated());
-        return redirect()->route('invites.show', $id);
+        $this->authorize('update', $invite);
+        $this->service->update($invite->id, $request->validated());
+        return redirect()->route('invites.show', $invite->id);
     }
 
-    public function destroy(int $id)
+    public function destroy(Invite $invite)
     {
-        $this->service->delete($id);
+        $this->authorize('delete', $invite);
+        $this->service->delete($invite->id);
         return redirect()->route('invites.index');
     }
 
-    public function send(int $id)
+    public function massDestroy(Request $request)
     {
-        $this->service->send($id);
+        $ids = $request->input('ids', []);
+        foreach ($ids as $id) {
+            $invite = $this->service->find($id);
+            $this->authorize('delete', $invite);
+            $this->service->delete($id);
+        }
+        return redirect()->route('invites.index');
+    }
+
+    public function massSend(Request $request)
+    {
+        $ids = $request->input('ids', []);
+        foreach ($ids as $id) {
+            $invite = $this->service->find($id);
+            $this->authorize('update', $invite);
+            $this->service->send($invite->id);
+        }
+        return redirect()->route('invites.index');
+    }
+
+    public function send(Invite $invite)
+    {
+        $this->authorize('update', $invite);
+        $this->service->send($invite->id);
         return redirect()->back();
     }
 }
